refactor(bolt): split verifier runtime into Bolt + Jolt audit tiers
Split the monolithic ~1.9k-LOC stages/common.rs into two files along an
explicit audit boundary:

  Tier A (Bolt verifier runtime):    stages/common.rs           1,265 LOC
    generic, protocol-agnostic plan structs, ValueStore,
    sumcheck driver loop, opening-equality interpreter,
    transcript helpers

  Tier B (audited Jolt verifier core): stages/jolt_relations.rs   638 LOC
    hand-written Jolt-specific verifier math: Stage 6/7
    evaluators, normalize_*_point, bytecode_gamma_powers,
    Stage67Bytecode* glue, polynomial-evaluation primitives

Tier B is the audit surface for Jolt-specific relation math; growth here
is now reviewed as a protocol-math decision rather than emitter LOC creep.
Tier A holds the generic Bolt scaffolding and is the long-term shrink
target as more helpers move into typed plan data driven from MLIR.

Wired through the artifact pipeline (verifier_runtime_modules now lists
both `common` and `jolt_relations`) and updated stage4/5/6 emitters to
split their import sites between super::common::{...} and
super::jolt_relations::{...}.

Per-tier hard LOC ceilings are enforced by verifier_cleanup.rs:
  BOLT_RUNTIME_BASELINE_LOC_CEILING       = 1_400  (current 1,265)
  JOLT_VERIFIER_CORE_BASELINE_LOC_CEILING =   700  (current 638)
  GENERATED_VERIFIER_TARGET_LOC           = 6_100  (current 6,002,
    bumped from 6,000 to absorb the import-split overhead)

GOAL.md gains an "Audit Tiers" section describing the A/B/C split and
records the post-split per-tier baseline. The pre-split "shared verifier
runtime" framing is retired.

cargo nextest -p bolt --test verifier_cleanup --test commitment_ir
cargo clippy -p bolt --all-targets -- -D warnings
cargo fmt --check
all green.
