+++ crates/bolt/src/protocols/jolt/emit/rust/stage5.rs
@@ -963,7 +963,8 @@ impl Stage5CpuProgram {
     }
 
     fn emit_verifier_imports() -> &'static str {
-        "use super::common::{batch_claims, eval_by_name, find_batch, find_plan, identity_polynomial_eval, indexed_evals_by_prefix_any, lt_polynomial_eval, normalize_instruction_read_raf_point, operand_polynomial_eval, reverse_slice, suffix_point};\n\
+        "use super::common::{batch_claims, eval_by_name, find_batch, find_plan, indexed_evals_by_prefix_any, reverse_slice, suffix_point};\n\
+         use super::jolt_relations::{identity_polynomial_eval, lt_polynomial_eval, normalize_instruction_read_raf_point, operand_polynomial_eval};\n\
          use jolt_field::{Field, Fr, RingCore};\n\
          use jolt_lookup_tables::LookupTableKind;\n\
          use jolt_poly::EqPolynomial;\n\
