+++ crates/jolt-verifier/src/stages/common.rs
@@ -6,8 +6,8 @@
 use std::fmt;
 use std::marker::PhantomData;
 
-use jolt_field::{Field, Fr, MulPow2, RingCore};
-use jolt_poly::{lagrange::lagrange_evals, EqPolynomial};
+use jolt_field::{Field, Fr, MulPow2};
+use jolt_poly::lagrange::lagrange_evals;
 use jolt_sumcheck::{
     CompressedLabeledRoundPoly, SumcheckClaim, SumcheckError, SumcheckProof, SumcheckVerifier,
 };
@@ -1127,14 +1127,6 @@ pub fn evaluate_field_expr<F: Field>(
     }
 }
 
-pub fn bytecode_gamma_powers(gamma: Fr) -> [Fr; 8] {
-    let mut powers = [Fr::from_u64(1); 8];
-    for index in 1..powers.len() {
-        powers[index] = powers[index - 1] * gamma;
-    }
-    powers
-}
-
 pub fn indexed_boolean_eq(index: usize, point: &[Fr]) -> Fr {
     point
         .iter()
@@ -1189,103 +1181,6 @@ pub fn suffix_point<'a, F: Field>(
         })
 }
 
-pub fn normalize_bytecode_read_raf_point<F: Field>(
-    point: &[F],
-    log_t: usize,
-    input: &'static str,
-) -> Result<Vec<F>, RuntimePlanError> {
-    let log_k = point
-        .len()
-        .checked_sub(log_t)
-        .ok_or(RuntimePlanError::InvalidInputLength {
-            input,
-            expected: log_t,
-            actual: point.len(),
-        })?;
-    let mut normalized = point.to_vec();
-    normalized[..log_k].reverse();
-    normalized[log_k..].reverse();
-    Ok(normalized)
-}
-
-pub fn normalize_instruction_read_raf_point<F: Field>(
-    point: &[F],
-    input: &'static str,
-) -> Result<Vec<F>, RuntimePlanError> {
-    const LOG_K: usize = 128;
-    if point.len() < LOG_K {
-        return Err(RuntimePlanError::InvalidInputLength {
-            input,
-            expected: LOG_K,
-            actual: point.len(),
-        });
-    }
-    let mut normalized = point.to_vec();
-    normalized[LOG_K..].reverse();
-    Ok(normalized)
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Stage67RelationSymbols {
-    pub hamming_booleanity_instance: &'static str,
-    pub booleanity_point: &'static str,
-    pub stage5_instruction_ra0: &'static str,
-    pub booleanity_combined_point: &'static str,
-    pub booleanity_gamma: &'static str,
-    pub booleanity_instruction_ra_prefix: &'static str,
-    pub booleanity_bytecode_ra_prefix: &'static str,
-    pub booleanity_ram_ra_prefix: &'static str,
-    pub hamming_weight_eval: &'static str,
-    pub hamming_lookup_output: &'static str,
-    pub ram_ra_virtual_cycle: &'static str,
-    pub ram_ra_virtual_eval_prefix: &'static str,
-    pub instruction_ra_virtual_cycle: &'static str,
-    pub instruction_ra_virtual_eval_prefix: &'static str,
-    pub instruction_ra_virtual_input_prefix: &'static str,
-    pub instruction_ra_virtual_gamma: &'static str,
-    pub inc_ram_stage2: &'static str,
-    pub inc_ram_stage4: &'static str,
-    pub inc_rd_stage4: &'static str,
-    pub inc_rd_stage5: &'static str,
-    pub inc_gamma: &'static str,
-    pub inc_ram_eval: &'static str,
-    pub inc_rd_eval: &'static str,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Stage67BytecodeSymbols {
-    pub point: &'static str,
-    pub gamma: &'static str,
-    pub bytecode_ra_eval_prefix: &'static str,
-    pub entries: &'static str,
-    pub entry_bytecode_index: &'static str,
-    pub stage_gammas: [&'static str; 5],
-    pub stage_cycle_points: [&'static str; 5],
-    pub stage4_register_point: &'static str,
-    pub stage5_register_point: &'static str,
-    pub entry_rd: &'static str,
-    pub entry_rs1: &'static str,
-    pub entry_rs2: &'static str,
-    pub entry_lookup_table: &'static str,
-}
-
-pub trait Stage67BytecodeEntry {
-    fn address(&self) -> Fr;
-    fn imm(&self) -> Fr;
-    fn circuit_flags(&self) -> &[bool; 14];
-    fn rd(&self) -> Option<usize>;
-    fn rs1(&self) -> Option<usize>;
-    fn rs2(&self) -> Option<usize>;
-    fn lookup_table(&self) -> Option<usize>;
-    fn is_interleaved(&self) -> bool;
-    fn is_branch(&self) -> bool;
-    fn left_is_rs1(&self) -> bool;
-    fn left_is_pc(&self) -> bool;
-    fn right_is_rs2(&self) -> bool;
-    fn right_is_imm(&self) -> bool;
-    fn is_noop(&self) -> bool;
-}
-
 pub fn store_scalar(store: &ValueStore<Fr>, symbol: &'static str) -> Result<Fr, RuntimePlanError> {
     store.scalar_or(symbol, |symbol| RuntimePlanError::MissingValue { symbol })
 }
@@ -1295,492 +1190,6 @@ pub fn store_point<'a>(
     symbol: &'static str,
 ) -> Result<&'a [Fr], RuntimePlanError> {
     store.point_or(symbol, |symbol| RuntimePlanError::MissingValue { symbol })
-}
-
-pub fn stage67_trace_rounds(
-    instance_results: &[SumcheckInstanceResultPlan],
-    symbols: &Stage67RelationSymbols,
-) -> Result<usize, RuntimePlanError> {
-    instance_results
-        .iter()
-        .find(|instance| instance.relation == RelationKind::Stage6HammingBooleanity)
-        .map(|instance| instance.num_rounds)
-        .ok_or(RuntimePlanError::MissingValue {
-            symbol: symbols.hamming_booleanity_instance,
-        })
-}
-
-pub fn expected_stage67_bytecode_read_raf<E: Stage67BytecodeEntry>(
-    entries: &[E],
-    entry_bytecode_index: usize,
-    num_lookup_tables: usize,
-    store: &ValueStore<Fr>,
-    evals: &[StageNamedEval<Fr>],
-    local_point: &[Fr],
-    log_t: usize,
-    symbols: &Stage67BytecodeSymbols,
-) -> Result<Fr, RuntimePlanError> {
-    let opening_point = normalize_bytecode_read_raf_point(local_point, log_t, symbols.point)?;
-    let log_k = opening_point.len() - log_t;
-    let (r_address_prime, r_cycle_prime) = opening_point.split_at(log_k);
-
-    let gamma = store_scalar(store, symbols.gamma)?;
-    let gamma_powers = bytecode_gamma_powers(gamma);
-    let int_eval = identity_polynomial_eval(r_address_prime);
-    let stage_value_evals = stage67_bytecode_stage_value_evals(
-        entries,
-        entry_bytecode_index,
-        num_lookup_tables,
-        store,
-        r_address_prime,
-        r_cycle_prime.len(),
-        symbols,
-    )?;
-    let stage_cycle_points =
-        stage67_bytecode_stage_cycle_points(store, r_cycle_prime.len(), symbols)?;
-    let int_contrib = [
-        gamma_powers[5] * int_eval,
-        Fr::from_u64(0),
-        gamma_powers[4] * int_eval,
-        Fr::from_u64(0),
-        Fr::from_u64(0),
-    ];
-
-    let mut val = Fr::from_u64(0);
-    for index in 0..stage_value_evals.len() {
-        val += (stage_value_evals[index] + int_contrib[index])
-            * EqPolynomial::<Fr>::mle(&stage_cycle_points[index], r_cycle_prime)
-            * gamma_powers[index];
-    }
-
-    let entry_bits = (0..log_k)
-        .map(|index| Fr::from_u64(((entry_bytecode_index >> (log_k - 1 - index)) & 1) as u64))
-        .collect::<Vec<_>>();
-    let zero_cycle = vec![Fr::from_u64(0); r_cycle_prime.len()];
-    let entry_contrib = gamma_powers[7]
-        * EqPolynomial::<Fr>::mle(&entry_bits, r_address_prime)
-        * EqPolynomial::<Fr>::mle(&zero_cycle, r_cycle_prime);
-    let bytecode_ra = indexed_evals_by_prefix_any(evals, symbols.bytecode_ra_eval_prefix)?
-        .into_iter()
-        .product::<Fr>();
-    Ok((val + entry_contrib) * bytecode_ra)
-}
-
-pub fn expected_stage67_booleanity(
-    store: &ValueStore<Fr>,
-    evals: &[StageNamedEval<Fr>],
-    local_point: &[Fr],
-    log_t: usize,
-    symbols: &Stage67RelationSymbols,
-) -> Result<Fr, RuntimePlanError> {
-    let log_k_chunk =
-        local_point
-            .len()
-            .checked_sub(log_t)
-            .ok_or(RuntimePlanError::InvalidInputLength {
-                input: symbols.booleanity_point,
-                expected: log_t,
-                actual: local_point.len(),
-            })?;
-    let stage5_point = store_point(store, symbols.stage5_instruction_ra0)?;
-    let stage5_address_len =
-        stage5_point
-            .len()
-            .checked_sub(log_t)
-            .ok_or(RuntimePlanError::InvalidInputLength {
-                input: symbols.stage5_instruction_ra0,
-                expected: log_t,
-                actual: stage5_point.len(),
-            })?;
-    if stage5_address_len < log_k_chunk {
-        return Err(RuntimePlanError::InvalidInputLength {
-            input: symbols.stage5_instruction_ra0,
-            expected: log_k_chunk + log_t,
-            actual: stage5_point.len(),
-        });
-    }
-
-    let mut stage5_addr = stage5_point[..stage5_address_len].to_vec();
-    stage5_addr.reverse();
-    let mut combined_r = stage5_addr[stage5_address_len - log_k_chunk..].to_vec();
-    combined_r.extend(stage5_point[stage5_address_len..].iter().rev().copied());
-    if combined_r.len() != local_point.len() {
-        return Err(RuntimePlanError::InvalidInputLength {
-            input: symbols.booleanity_combined_point,
-            expected: local_point.len(),
-            actual: combined_r.len(),
-        });
-    }
-    let mut verifier_point = combined_r[..log_k_chunk].to_vec();
-    verifier_point.reverse();
-    verifier_point.extend(combined_r[log_k_chunk..].iter().rev().copied());
-    let eq_eval = EqPolynomial::<Fr>::mle(local_point, &verifier_point);
-
-    let gamma = store_scalar(store, symbols.booleanity_gamma)?;
-    let gamma_sq = gamma.square();
-    let mut gamma_power = Fr::from_u64(1);
-    let mut booleanity = Fr::from_u64(0);
-    for ra in stage67_booleanity_evals(evals, symbols)? {
-        booleanity += gamma_power * (ra.square() - ra);
-        gamma_power *= gamma_sq;
-    }
-    Ok(eq_eval * booleanity)
-}
-
-pub fn expected_stage67_hamming_booleanity(
-    store: &ValueStore<Fr>,
-    evals: &[StageNamedEval<Fr>],
-    local_point: &[Fr],
-    symbols: &Stage67RelationSymbols,
-) -> Result<Fr, RuntimePlanError> {
-    let hamming = eval_by_name(evals, symbols.hamming_weight_eval)?;
-    let lookup_output_point = reverse_slice(store_point(store, symbols.hamming_lookup_output)?);
-    if lookup_output_point.len() != local_point.len() {
-        return Err(RuntimePlanError::InvalidInputLength {
-            input: symbols.hamming_lookup_output,
-            expected: local_point.len(),
-            actual: lookup_output_point.len(),
-        });
-    }
-    let eq_eval = EqPolynomial::<Fr>::mle(local_point, &lookup_output_point);
-    Ok((hamming.square() - hamming) * eq_eval)
-}
-
-pub fn expected_stage67_ram_ra_virtual(
-    store: &ValueStore<Fr>,
-    evals: &[StageNamedEval<Fr>],
-    local_point: &[Fr],
-    symbols: &Stage67RelationSymbols,
-) -> Result<Fr, RuntimePlanError> {
-    let r_cycle_reduced = reverse_slice(local_point);
-    let r_cycle = suffix_point(
-        store_point(store, symbols.ram_ra_virtual_cycle)?,
-        r_cycle_reduced.len(),
-        symbols.ram_ra_virtual_cycle,
-    )?;
-    let eq_eval = EqPolynomial::<Fr>::mle(r_cycle, &r_cycle_reduced);
-    let ram_ra = indexed_evals_by_prefix_any(evals, symbols.ram_ra_virtual_eval_prefix)?
-        .into_iter()
-        .product::<Fr>();
-    Ok(eq_eval * ram_ra)
-}
-
-pub fn expected_stage67_instruction_ra_virtual(
-    opening_inputs: &[OpeningInputPlan],
-    store: &ValueStore<Fr>,
-    evals: &[StageNamedEval<Fr>],
-    local_point: &[Fr],
-    symbols: &Stage67RelationSymbols,
-) -> Result<Fr, RuntimePlanError> {
-    let r_cycle_reduced = reverse_slice(local_point);
-    let r_cycle = suffix_point(
-        store_point(store, symbols.instruction_ra_virtual_cycle)?,
-        r_cycle_reduced.len(),
-        symbols.instruction_ra_virtual_cycle,
-    )?;
-    let eq_eval = EqPolynomial::<Fr>::mle(r_cycle, &r_cycle_reduced);
-    let committed_ra =
-        indexed_evals_by_prefix_any(evals, symbols.instruction_ra_virtual_eval_prefix)?;
-    let virtual_count = opening_inputs
-        .iter()
-        .filter(|input| {
-            input
-                .symbol
-                .starts_with(symbols.instruction_ra_virtual_input_prefix)
-        })
-        .count();
-    if virtual_count == 0 || committed_ra.len() % virtual_count != 0 {
-        return Err(RuntimePlanError::InvalidInputLength {
-            input: symbols.instruction_ra_virtual_eval_prefix,
-            expected: virtual_count,
-            actual: committed_ra.len(),
-        });
-    }
-    let committed_per_virtual = committed_ra.len() / virtual_count;
-    let gamma = store_scalar(store, symbols.instruction_ra_virtual_gamma)?;
-    let mut gamma_power = Fr::from_u64(1);
-    let mut value = Fr::from_u64(0);
-    for chunk in committed_ra.chunks(committed_per_virtual) {
-        value += gamma_power * chunk.iter().copied().product::<Fr>();
-        gamma_power *= gamma;
-    }
-    Ok(eq_eval * value)
-}
-
-pub fn expected_stage67_inc_claim_reduction(
-    store: &ValueStore<Fr>,
-    evals: &[StageNamedEval<Fr>],
-    local_point: &[Fr],
-    symbols: &Stage67RelationSymbols,
-) -> Result<Fr, RuntimePlanError> {
-    let r_cycle_reduced = reverse_slice(local_point);
-    let ram_inc_stage2 = suffix_point(
-        store_point(store, symbols.inc_ram_stage2)?,
-        r_cycle_reduced.len(),
-        symbols.inc_ram_stage2,
-    )?;
-    let ram_inc_stage4 = suffix_point(
-        store_point(store, symbols.inc_ram_stage4)?,
-        r_cycle_reduced.len(),
-        symbols.inc_ram_stage4,
-    )?;
-    let rd_inc_stage4 = suffix_point(
-        store_point(store, symbols.inc_rd_stage4)?,
-        r_cycle_reduced.len(),
-        symbols.inc_rd_stage4,
-    )?;
-    let rd_inc_stage5 = suffix_point(
-        store_point(store, symbols.inc_rd_stage5)?,
-        r_cycle_reduced.len(),
-        symbols.inc_rd_stage5,
-    )?;
-    let gamma = store_scalar(store, symbols.inc_gamma)?;
-    let eq_ram_combined = EqPolynomial::<Fr>::mle(ram_inc_stage2, &r_cycle_reduced)
-        + gamma * EqPolynomial::<Fr>::mle(ram_inc_stage4, &r_cycle_reduced);
-    let eq_rd_combined = EqPolynomial::<Fr>::mle(rd_inc_stage4, &r_cycle_reduced)
-        + gamma * EqPolynomial::<Fr>::mle(rd_inc_stage5, &r_cycle_reduced);
-    let ram_inc = eval_by_name(evals, symbols.inc_ram_eval)?;
-    let rd_inc = eval_by_name(evals, symbols.inc_rd_eval)?;
-    Ok(ram_inc * eq_ram_combined + gamma.square() * rd_inc * eq_rd_combined)
-}
-
-fn stage67_booleanity_evals(
-    evals: &[StageNamedEval<Fr>],
-    symbols: &Stage67RelationSymbols,
-) -> Result<Vec<Fr>, RuntimePlanError> {
-    let mut values = indexed_evals_by_prefix_any(evals, symbols.booleanity_instruction_ra_prefix)?;
-    values.extend(indexed_evals_by_prefix_any(
-        evals,
-        symbols.booleanity_bytecode_ra_prefix,
-    )?);
-    values.extend(indexed_evals_by_prefix_any(
-        evals,
-        symbols.booleanity_ram_ra_prefix,
-    )?);
-    Ok(values)
-}
-
-fn stage67_bytecode_stage_cycle_points(
-    store: &ValueStore<Fr>,
-    log_t: usize,
-    symbols: &Stage67BytecodeSymbols,
-) -> Result<[Vec<Fr>; 5], RuntimePlanError> {
-    let point = |index| {
-        let symbol = symbols.stage_cycle_points[index];
-        suffix_point(store_point(store, symbol)?, log_t, symbol).map(|point| point.to_vec())
-    };
-    Ok([point(0)?, point(1)?, point(2)?, point(3)?, point(4)?])
-}
-
-fn stage67_bytecode_stage_value_evals<E: Stage67BytecodeEntry>(
-    entries: &[E],
-    entry_bytecode_index: usize,
-    num_lookup_tables: usize,
-    store: &ValueStore<Fr>,
-    r_address: &[Fr],
-    log_t: usize,
-    symbols: &Stage67BytecodeSymbols,
-) -> Result<[Fr; 5], RuntimePlanError> {
-    let expected_len =
-        1usize
-            .checked_shl(r_address.len() as u32)
-            .ok_or(RuntimePlanError::InvalidInputLength {
-                input: symbols.entries,
-                expected: usize::BITS as usize,
-                actual: r_address.len(),
-            })?;
-    if entries.len() != expected_len {
-        return Err(RuntimePlanError::InvalidInputLength {
-            input: symbols.entries,
-            expected: expected_len,
-            actual: entries.len(),
-        });
-    }
-    if entry_bytecode_index >= expected_len {
-        return Err(RuntimePlanError::InvalidInputLength {
-            input: symbols.entry_bytecode_index,
-            expected: expected_len,
-            actual: entry_bytecode_index + 1,
-        });
-    }
-
-    let stage1_gamma_powers = field_powers(store_scalar(store, symbols.stage_gammas[0])?, 16);
-    let stage2_gamma_powers = field_powers(store_scalar(store, symbols.stage_gammas[1])?, 4);
-    let stage3_gamma_powers = field_powers(store_scalar(store, symbols.stage_gammas[2])?, 9);
-    let stage4_gamma_powers = field_powers(store_scalar(store, symbols.stage_gammas[3])?, 3);
-    let stage5_gamma_powers = field_powers(
-        store_scalar(store, symbols.stage_gammas[4])?,
-        num_lookup_tables + 2,
-    );
-
-    let stage4_register_point =
-        stage67_register_prefix_point(store, symbols.stage4_register_point, log_t)?;
-    let stage5_register_point =
-        stage67_register_prefix_point(store, symbols.stage5_register_point, log_t)?;
-
-    let mut evals = [Fr::from_u64(0); 5];
-    for (index, entry) in entries.iter().enumerate() {
-        let eq = indexed_boolean_eq(index, r_address);
-        let values = stage67_bytecode_entry_stage_values(
-            entry,
-            num_lookup_tables,
-            stage4_register_point,
-            stage5_register_point,
-            &stage1_gamma_powers,
-            &stage2_gamma_powers,
-            &stage3_gamma_powers,
-            &stage4_gamma_powers,
-            &stage5_gamma_powers,
-            symbols,
-        )?;
-        for stage in 0..evals.len() {
-            evals[stage] += eq * values[stage];
-        }
-    }
-    Ok(evals)
-}
-
-fn stage67_bytecode_entry_stage_values<E: Stage67BytecodeEntry>(
-    entry: &E,
-    num_lookup_tables: usize,
-    stage4_register_point: &[Fr],
-    stage5_register_point: &[Fr],
-    stage1_gamma_powers: &[Fr],
-    stage2_gamma_powers: &[Fr],
-    stage3_gamma_powers: &[Fr],
-    stage4_gamma_powers: &[Fr],
-    stage5_gamma_powers: &[Fr],
-    symbols: &Stage67BytecodeSymbols,
-) -> Result<[Fr; 5], RuntimePlanError> {
-    let flags = entry.circuit_flags();
-    let mut stage1 = entry.address() + entry.imm() * stage1_gamma_powers[1];
-    for (flag, gamma) in flags.iter().zip(stage1_gamma_powers.iter().skip(2)) {
-        if *flag {
-            stage1 += *gamma;
-        }
-    }
-
-    let mut stage2 = Fr::from_u64(0);
-    if flags[5] {
-        stage2 += stage2_gamma_powers[0];
-    }
-    if entry.is_branch() {
-        stage2 += stage2_gamma_powers[1];
-    }
-    if flags[6] {
-        stage2 += stage2_gamma_powers[2];
-    }
-    if flags[7] {
-        stage2 += stage2_gamma_powers[3];
-    }
-
-    let mut stage3 = entry.imm() + entry.address() * stage3_gamma_powers[1];
-    if entry.left_is_rs1() {
-        stage3 += stage3_gamma_powers[2];
-    }
-    if entry.left_is_pc() {
-        stage3 += stage3_gamma_powers[3];
-    }
-    if entry.right_is_rs2() {
-        stage3 += stage3_gamma_powers[4];
-    }
-    if entry.right_is_imm() {
-        stage3 += stage3_gamma_powers[5];
-    }
-    if entry.is_noop() {
-        stage3 += stage3_gamma_powers[6];
-    }
-    if flags[7] {
-        stage3 += stage3_gamma_powers[7];
-    }
-    if flags[12] {
-        stage3 += stage3_gamma_powers[8];
-    }
-
-    let stage4 = stage67_register_eq(entry.rd(), stage4_register_point, symbols.entry_rd)?
-        * stage4_gamma_powers[0]
-        + stage67_register_eq(entry.rs1(), stage4_register_point, symbols.entry_rs1)?
-            * stage4_gamma_powers[1]
-        + stage67_register_eq(entry.rs2(), stage4_register_point, symbols.entry_rs2)?
-            * stage4_gamma_powers[2];
-
-    let mut stage5 = stage67_register_eq(entry.rd(), stage5_register_point, symbols.entry_rd)?
-        * stage5_gamma_powers[0];
-    if !entry.is_interleaved() {
-        stage5 += stage5_gamma_powers[1];
-    }
-    if let Some(table) = entry.lookup_table() {
-        if table >= num_lookup_tables {
-            return Err(RuntimePlanError::InvalidInputLength {
-                input: symbols.entry_lookup_table,
-                expected: num_lookup_tables,
-                actual: table + 1,
-            });
-        }
-        stage5 += stage5_gamma_powers[2 + table];
-    }
-
-    Ok([stage1, stage2, stage3, stage4, stage5])
-}
-
-fn stage67_register_eq(
-    index: Option<usize>,
-    point: &[Fr],
-    input: &'static str,
-) -> Result<Fr, RuntimePlanError> {
-    let Some(index) = index else {
-        return Ok(Fr::from_u64(0));
-    };
-    let register_count =
-        1usize
-            .checked_shl(point.len() as u32)
-            .ok_or(RuntimePlanError::InvalidInputLength {
-                input,
-                expected: usize::BITS as usize,
-                actual: point.len(),
-            })?;
-    if index >= register_count {
-        return Err(RuntimePlanError::InvalidInputLength {
-            input,
-            expected: register_count,
-            actual: index + 1,
-        });
-    }
-    Ok(indexed_boolean_eq(index, point))
-}
-
-fn stage67_register_prefix_point<'a>(
-    store: &'a ValueStore<Fr>,
-    symbol: &'static str,
-    log_t: usize,
-) -> Result<&'a [Fr], RuntimePlanError> {
-    let point = store_point(store, symbol)?;
-    let register_len =
-        point
-            .len()
-            .checked_sub(log_t)
-            .ok_or(RuntimePlanError::InvalidInputLength {
-                input: symbol,
-                expected: log_t,
-                actual: point.len(),
-            })?;
-    prefix_point(point, register_len, symbol)
-}
-
-pub fn operand_polynomial_eval(point: &[Fr], left: bool) -> Fr {
-    let stride_offset = usize::from(!left);
-    let operand_bits = point.len() / 2;
-    (0..operand_bits)
-        .map(|index| point[2 * index + stride_offset].mul_pow_2(operand_bits - 1 - index))
-        .sum()
-}
-
-pub fn identity_polynomial_eval(point: &[Fr]) -> Fr {
-    point
-        .iter()
-        .enumerate()
-        .map(|(index, value)| value.mul_pow_2(point.len() - 1 - index))
-        .sum()
 }
 
 pub fn append_labeled_scalar<T>(transcript: &mut T, label: &'static str, scalar: &Fr)
@@ -1836,16 +1245,6 @@ where
         }
     }
     Ok(())
-}
-
-pub fn lt_polynomial_eval(x: &[Fr], y: &[Fr]) -> Fr {
-    let mut lt_eval = Fr::from_u64(0);
-    let mut eq_term = Fr::from_u64(1);
-    for (x_i, y_i) in x.iter().zip(y.iter()) {
-        lt_eval += (Fr::from_u64(1) - *x_i) * *y_i * eq_term;
-        eq_term *= Fr::from_u64(1) - *x_i - *y_i + *x_i * *y_i + *x_i * *y_i;
-    }
-    lt_eval
 }
 
 pub fn pow_field<F: Field>(base: F, mut exponent: usize) -> F {
